Moved Serializer to a new folder
