+++ tests/Mongolid/Serializer/SerializerTest.php
@@ -1,0 +1,11 @@
+<?php
+namespace Mongolid;
+
+use PHPUnit_Framework_TestCase as TestCase;
+
+/**
+ * Test case for Serializer class
+ */
+class SerializerTest extends TestCase
+{
+}
